refactor: Move merge stategy into own file
